🔀 Refactor to enable additional inventory APIs
