+++ src/classes/InventoryGetter.ts
@@ -2,13 +2,14 @@ import SteamID from 'steamid';
 import { EconItem } from '@tf2autobot/tradeoffer-manager';
 
 import Bot from './Bot';
+import InventoryAPI from './InventoryAPI';
 import SteamApis from './SteamApis';
 
 export default class InventoryGetter {
-    private readonly steamApis: SteamApis;
+    private readonly inventoryAPIs: InventoryAPI[];
 
     constructor(private readonly bot: Bot) {
-        this.steamApis = new SteamApis(this.bot);
+        this.inventoryAPIs = [new SteamApis(this.bot)];
     }
 
     getUserInventoryContents(
@@ -18,8 +19,16 @@ export default class InventoryGetter {
         tradeableOnly: boolean,
         callback: (err?: Error, inventory?: EconItem[], currencies?: EconItem[]) => void
     ): void {
-        if (!this.bot.options.steamApis.enable || steamID.toString() == this.bot.client.steamID.getSteamID64())
+        if (steamID.toString() == this.bot.client.steamID.getSteamID64()) {
             this.bot.manager.getUserInventoryContents(steamID, appid, contextid, tradeableOnly, callback);
-        else this.steamApis.getUserInventoryContents(steamID, appid, contextid, tradeableOnly, callback);
+            return;
+        }
+        for (const inventoryAPI of this.inventoryAPIs) {
+            if (inventoryAPI.isEnabled()) {
+                inventoryAPI.getUserInventoryContents(steamID, appid, contextid, tradeableOnly, callback);
+                return;
+            }
+        }
+        this.bot.manager.getUserInventoryContents(steamID, appid, contextid, tradeableOnly, callback);
     }
 }
